AFL: Automatically resume from corpus if present
AFL will refuse to start if an input directory is given and the output
directory indicates the fuzzer has already run for at least half an
hour. If the output directory indicates the fuzzer has run before, then
don't pass the input directory, so that AFL will resume from where it
left off.